VER-585 [AL] helpline-frontend: (FE) Create new static contact page for 'Child Benefits'
