+++ app/uk/gov/hmrc/helplinefrontend/views/ChildBenefit.scala.html
@@ -1,0 +1,43 @@
+@*
+ * Copyright 2021 HM Revenue & Customs
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *@
+
+@import uk.gov.hmrc.helplinefrontend.config.AppConfig
+
+@this(layout: Layout)
+
+@()(implicit request: Request[_], messages: Messages, appConfig: AppConfig)
+
+@layout(pageTitle = Some(s"${messages("child-benefit.title")} ${messages("title.suffix")}")) {
+ <h1 class="govuk-heading-xl ">
+ @messages("child-benefit.title")
+ </h1>
+<dl>
+   <dt class="govuk-body govuk-!-margin-bottom-0">@messages("label.telephone"):</dt>
+   <dd class="govuk-heading-l govuk-!-margin-top-0 govuk-!-padding-top-0 govuk-!-font-weight-bold govuk-!-margin-left-0">@messages("child-benefit.telephone")</dd>
+  </dl>
+  <dl>
+   <dt class="govuk-body govuk-!-margin-bottom-0">@messages("label.outside-uk"):</dt>
+   <dd class="govuk-body govuk-!-font-weight-bold govuk-!-margin-left-0">@messages("child-benefit.telephone.international")6</dd>
+  </dl>
+
+  <p class="govuk-body govuk-hint govuk-!-margin-bottom-0">@messages("label.monday-friday"): @messages("child-benefit.hours.monday-friday")</p>
+  <p class="govuk-body govuk-hint">@messages("child-benefit.closed")</p>
+
+  <div class="box--shaded">
+   <p class="govuk-body">@Html(messages("get-help", "https://www.gov.uk/get-help-hmrc-extra-support"))</p>
+  </div>
+  <p class="govuk-body govuk-!-margin-bottom-9">@Html(messages("call-charges", "https://www.gov.uk/call-charges"))</p>
+}
