+++ conf/application.conf
@@ -107,6 +107,6 @@ google-analytics {
 play.i18n.langs = ["en", "cy"]
 
 # Change this value to true to enable Welsh translations to be loaded from messages.cy, and to display the language toggle
-features.welsh-language-support = false
+features.welsh-language-support = true
