+++ conf/messages.cy
@@ -1,3 +1,21 @@
-service.name = helpline-frontend
-service.homePageUrl = /helpline-frontend
-service.text = Dyma'ch gwasanaeth newydd
+service.name=
+title.suffix=- GOV.UK
+
+label.telephone=Ffôn
+label.outside-uk=Y tu allan i’r DU
+label.monday-friday=Dydd Llun – Dydd Gwener
+call-charges=<a class="govuk-link" href="{0}">Find out about call charges</a>.
+get-help=Get help <a class="govuk-link" href="{0}">dealing with HMRC if you have additional needs</a>.
+
+child-benefit.h1=Call the Child Benefits helpline
+child-benefit.telephone=0300 200 3100
+child-benefit.telephone.international=+44 161 210 3086
+child-benefit.hours.monday-friday=8:30am - 5pm
+child-benefit.closed=Ar gau ar benwythnosau a gwyliau banc.
+
+
+
+
+
+
+
